Move scope-related logic into its own directory
