+++ packages/arbor-store/src/index.ts
@@ -8,7 +8,7 @@ export { NodeHandler } from "./handlers/NodeHandler"
 export { ImmutableArbor } from "./ImmutableArbor"
 export * from "./MutationEngine"
 export { Path } from "./path"
-export * from "./ScopedStore"
+export * from "./scoping/ScopedStore"
 export { Subscriptions } from "./Subscriptions"
 export type * from "./types"
 export * from "./utilities"
